move chat stateless components to components folder
