+++ client/src/containers/Chat/Chat.js
@@ -13,9 +13,9 @@ import {
 } from "../../actions/chatActions";
 import { _loadProfile } from "../../actions/userLoadActions";
 import Layout from "../Layout";
-import AddMessage from "./AddMessage";
-import UserList from "./UserList";
-import MessageList from "./MessageList";
+import AddMessage from "../../components/Chat/AddMessage";
+import UserList from "../../components/Chat/UserList";
+import MessageList from "../../components/Chat/MessageList";
 
 class Chat extends Component {
   componentDidMount = async () => {
@@ -26,13 +26,15 @@ class Chat extends Component {
       await this.props._loadChatSessions(localStorage.getItem("token"));
       await this.props._loadProfile(localStorage.getItem("token"));
     } else {
-
       if (this.props.dealItem) {
-        await this.props.history.push(`/SignIn?redirect=feed/deals/${this.props.dealItem.deal_id}/${this.props.dealItem.deal_name}`);
+        await this.props.history.push(
+          `/SignIn?redirect=feed/deals/${this.props.dealItem.deal_id}/${
+            this.props.dealItem.deal_name
+          }`
+        );
       } else {
         this.props.history.push("/");
       }
-
     }
   };
 
@@ -43,7 +45,6 @@ class Chat extends Component {
     let seller_id = this.props.selected_chat_session[0].seller_id;
     let user_id = this.props.user_info[0].id;
     let recipientEmailUser_id;
-
 
     if (buyer_id === user_id) {
       recipientEmailUser_id = seller_id;
@@ -67,7 +68,6 @@ class Chat extends Component {
   };
 
   deleteChatSession = async event => {
-
     let chat_session_id = this.props.selected_chat_session[0].chat_session_id;
 
     await this.props._deleteChatSession(
@@ -76,16 +76,17 @@ class Chat extends Component {
     );
 
     await this.props._loadChatSessions(localStorage.getItem("token"));
-
   };
 
   componentDidUpdate() {
     //scroll to bottom of message list
-    if (this.props.chat_sessions.length > 0 && this.props.selected_chat_session.length > 0) {
+    if (
+      this.props.chat_sessions.length > 0 &&
+      this.props.selected_chat_session.length > 0
+    ) {
       const objDiv = document.getElementById("chat-messages");
       objDiv.scrollTop = objDiv.scrollHeight;
     }
-
   }
 
   render() {
@@ -106,55 +107,67 @@ class Chat extends Component {
         <Layout>
           <div className="chat-sessions">
             <section className="chat-session-left">
-            {chat_sessions.length > 0 ?
-              <UserList
-                _loadUsersList={_loadChatSessions}
-                usersList={chat_sessions}
-                _fetchMessagesList={_loadChatMessages}
-                userInfo={user_info}
-                chatSessionInfo={selected_chat_session}
-                {...history}
-              /> :
-              <div id="empty-chat-session">
-                <div>
-                  <div><i className="far fa-4x fa-envelope"></i></div>
-                  <strong>You haven't started any conversation yet.</strong>
-                  <br/>
-                  <strong> Check out these <Link style={{textDecoration: "none"}} to={"/"}>DEALS.</Link></strong>
+              {chat_sessions.length > 0 ? (
+                <UserList
+                  _loadUsersList={_loadChatSessions}
+                  usersList={chat_sessions}
+                  _fetchMessagesList={_loadChatMessages}
+                  userInfo={user_info}
+                  chatSessionInfo={selected_chat_session}
+                  {...history}
+                />
+              ) : (
+                <div id="empty-chat-session">
+                  <div>
+                    <div>
+                      <i className="far fa-4x fa-envelope" />
+                    </div>
+                    <strong>You haven't started any conversation yet.</strong>
+                    <br />
+                    <strong>
+                      {" "}
+                      Check out these{" "}
+                      <Link style={{ textDecoration: "none" }} to={"/"}>
+                        DEALS.
+                      </Link>
+                    </strong>
+                  </div>
                 </div>
-              </div>
-            }
+              )}
             </section>
 
             <section className="chat-session-right">
-            {chat_sessions.length > 0 && selected_chat_session.length > 0 ?
-            <div>
-              <div>
-                <MessageList
-                  messagesList={chat_messages}
-                  chatSessionInfo={selected_chat_session}
-                  userInfo={user_info}
-                  _deleteChatSession={this.deleteChatSession}
-                />
-              </div>
-              <div>
-                <hr />
+              {chat_sessions.length > 0 && selected_chat_session.length > 0 ? (
                 <div>
-                  <AddMessage
-                    _createMessage={this.addMessage}
-                    handleChatMessage={onMessageEdit}
-                    message={chatMessageValue}
-                  />
+                  <div>
+                    <MessageList
+                      messagesList={chat_messages}
+                      chatSessionInfo={selected_chat_session}
+                      userInfo={user_info}
+                      _deleteChatSession={this.deleteChatSession}
+                    />
+                  </div>
+                  <div>
+                    <hr />
+                    <div>
+                      <AddMessage
+                        _createMessage={this.addMessage}
+                        handleChatMessage={onMessageEdit}
+                        message={chatMessageValue}
+                      />
+                    </div>
+                  </div>
                 </div>
-              </div>
-            </div> :
-            <div id="select-chat-prompt">
-              <div>
-                <div><i className="fas fa-7x fa-comments"></i></div>
-                <strong>Select a conversation to start the chat</strong>
-              </div>
-            </div>
-            }
+              ) : (
+                <div id="select-chat-prompt">
+                  <div>
+                    <div>
+                      <i className="fas fa-7x fa-comments" />
+                    </div>
+                    <strong>Select a conversation to start the chat</strong>
+                  </div>
+                </div>
+              )}
             </section>
           </div>
         </Layout>
